optimizations with helper functions for image padding
